:sparkles: Creacion de instancias

+++ co.edu.uniquindio.estudiante/src/main/java/co/edu/uniquindio/estudiante/model/Estudiante.java
@@ -2,12 +2,12 @@ package co.edu.uniquindio.estudiante.model;
 public class Estudiante
 {
     private String nombre;
-    private int edad;
+    private double edad;
     private String correo;
     private int semestre;
-    private float nota1;
-    private float nota2;
-    private float nota3;
+    private double nota1;
+    private double nota2;
+    private double nota3;
 
     public String getNombre() {
         return nombre;
@@ -17,11 +17,11 @@ public class Estudiante
         this.nombre = nombre;
     }
 
-    public int getEdad() {
+    public double getEdad() {
         return edad;
     }
 
-    public void setEdad(int edad) {
+    public void setEdad(double edad) {
         this.edad = edad;
     }
 
@@ -41,27 +41,27 @@ public class Estudiante
         this.semestre = semestre;
     }
 
-    public float getNota1() {
+    public double getNota1() {
         return nota1;
     }
 
-    public void setNota1(float nota1) {
+    public void setNota1(double nota1) {
         this.nota1 = nota1;
     }
 
-    public float getNota2() {
+    public double getNota2() {
         return nota2;
     }
 
-    public void setNota2(float nota2) {
+    public void setNota2(double nota2) {
         this.nota2 = nota2;
     }
 
-    public float getNota3() {
+    public double getNota3() {
         return nota3;
     }
 
-    public void setNota3(float nota3) {
+    public void setNota3(double nota3) {
         this.nota3 = nota3;
     }
 }
